remove let's encrypt integration as it's been busted for a while
