+++ src/lib.rs
@@ -2,4 +2,3 @@ pub mod common;
 pub mod crtsh;
 pub mod ct;
 pub mod google;
-pub mod letsencrypt;
